% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debbie.R
\name{debPkgAvailable}
\alias{debPkgAvailable}
\title{Determine Whether an R Package Exists in Debian Sources Repository}
\usage{
debPkgAvailable(package, deb_mirror, sources_url)
}
\arguments{
\item{package}{A character string describing an R package, for which a search of the Debian package repository will be performed.}

\item{deb_mirror}{A character string which represents a valid URL to a Debian package mirror's R package tree. Default is to use http://deb.debian.org/debian/pool/main/r.}

\item{sources_url}{A character string which represents a valid URL to a Debian sources API. Default is https://sources.debian.org/api/src.}
}
\description{
This function attempts to query the Debian sources repository API for
an R package, provided as a string. The function returns a list with the
API's response as well as a value of \code{TRUE} if found, and \code{FALSE}
otherwise.
}
