% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debbie.R
\name{install_deb}
\alias{install_deb}
\title{Install an R package from the Debian Package Repository}
\usage{
install_deb(package = NULL, url = NULL, path = "/tmp",
  clean = TRUE, ...)
}
\arguments{
\item{package}{A character string describing an R package, for which a search of the Debian package repository will be performed.}

\item{url}{A character string which represents a valid URL to an R package.`}

\item{path}{A character string describing the intended destination directory of the package.}

\item{clean}{Logical. A Boolean flag indicating whether the non-package directories extracted from the archive should be deleted.}

\item{...}{Arguments to be passed on to `install.packages`.}
}
\description{
This function attempts to retrieve a Debian binary package corresponding
to a (source) R package, which is then extracted and installed, without
having to configure or edit the apt sources.list file.
}
\details{
This can be useful if you're in a hurry, or want to simplify the install
process for users who have limited familiarity with system configuration
but wish to avoid the time required to retrieve and compile R packages
for Debian or Ubuntu from CRAN or GitHub.
}
\keyword{Debian}
\keyword{binary}
\keyword{install}
\keyword{packages}
