% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debbie.R
\name{retrievePackage}
\alias{retrievePackage}
\title{Retrieve an R package binary from Debian's package repository}
\usage{
retrievePackage(url, path = tempdir())
}
\arguments{
\item{url}{A character string which represents a valid URL to an R package.}

\item{path}{A character string describing the path where the package should be downloaded. Defaults to \code{tempdir()}.}
}
\description{
This function attempts to download an R package compiled for
Debian (which may work on hosts running Ubuntu also) using
curl to the directory specified in \code{path}.
}
\keyword{Debian}
\keyword{binary}
\keyword{download}
\keyword{package}
