% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debbie.R
\name{unpackPackage}
\alias{unpackPackage}
\title{Extract a Debian archive containing an R package binary}
\usage{
unpackPackage(pkg_path = tempdir(), pkg_file, dest_path = tempdir())
}
\arguments{
\item{pkg_path}{A character string describing the location of the compressed package.}

\item{pkg_file}{A character string describing the filename of the Debian package.}

\item{dest_path}{A character string describing the intended destination directory of the package.}
}
\description{
This function uses \code{untar} to extract the R package from within a Debian
archive.
}
\keyword{Debian}
\keyword{binary}
\keyword{extract}
\keyword{untar}
