% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debbie-package.R
\docType{package}
\name{debbie-package}
\alias{debbie}
\alias{debbie-package}
\title{Retrieve and Install Debian Binary Versions of R Packages}
\description{
Coming soon
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rpkyle/debbie}
  \item Report bugs at \url{https://github.com/rpkyle/debbie/issues}
}

}
\author{
\strong{Maintainer}: Ryan Patrick Kyle \email{ryan@plot.ly}

}
