% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debbie.R
\name{install_deb}
\alias{install_deb}
\title{Install an R package from the Debian Package Repository}
\usage{
install_deb(package = NULL, pkg_ver = NULL,
  mirror = "http://deb.debian.org/debian/pool/main/r",
  sources_url = "https://sources.debian.org/api/src", release = "sid",
  download_path = tempdir(), clean = FALSE, opts = c("--no-docs",
  "--no-multiarch", "--no-demo"), echo = FALSE, show = TRUE,
  fail_on_status = TRUE, ...)
}
\arguments{
\item{package}{A character string describing an R package, for which a search of the Debian package repository will be performed.}

\item{pkg_ver}{A character string describing a particular version of an R package, for which a search will be performed.}

\item{mirror}{A character string which represents a valid URL to a Debian package mirror's R package tree. Default is to use http://deb.debian.org/debian/pool/main/r.}

\item{sources_url}{A character string which represents a valid URL to a Debian sources API. Default is https://sources.debian.org/api/src.}

\item{release}{A character string describing the desired Debian release code name, default is \code{sid}.}

\item{clean}{Logical. A Boolean flag indicating whether the non-package directories extracted from the archive should be deleted.}

\item{opts}{A vector of character strings containing command line arguments for \code{INSTALL}, used when installing the downloaded package. Default is \code{--no-docs}, \code{--no-multiarch}, \code{--no-demo}.}

\item{echo}{Logical. A Boolean flag passed to \code{callr::rcmd} which indicates whether the complete command should be echoed to the R console.}

\item{show}{Logical. A Boolean flag passed to \code{callr::rcmd} which indicates whether the standard output of the \code{INSTALL} command run by \code{callr::rcmd} should be displayed while the process is running.}

\item{fail_on_status}{Logical. A Boolean flag passed to \code{callr::rcmd} which controls whether an error should be thrown if the underlying process terminates with a status code other than 0. Default is \code{TRUE}.}

\item{...}{Arguments to be passed on to \code{remotes::install_deps}.}

\item{pkg_path}{A character string describing the intended destination directory of the package when downloaded. Default is the session temporary directory given by \code{tempdir()}.}
}
\description{
This function attempts to retrieve a Debian binary package corresponding
to a (source) R package, which is then extracted and installed, without
having to configure or edit the apt sources.list file.
}
\details{
This can be useful if you're in a hurry, or want to simplify the install
process for users who have limited familiarity with system configuration
but wish to avoid the time required to retrieve and compile R packages
for Debian or Ubuntu from CRAN or GitHub.
}
\keyword{Debian}
\keyword{binary}
\keyword{install}
\keyword{packages}
