% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debbie.R
\name{unpackPackage}
\alias{unpackPackage}
\title{Extract a Debian archive containing an R package binary}
\usage{
unpackPackage(pkg_path, dest_path, clean = TRUE)
}
\arguments{
\item{pkg_path}{A character string describing the location of the compressed package.}

\item{dest_path}{A character string describing the intended destination directory of the package.}

\item{clean}{Logical. A Boolean flag indicating whether the non-package directories extracted from the archive should be deleted.}
}
\description{
This function uses `untar` to extract the R package from within a Debian
archive, and, if `clean` is `TRUE`, deletes all non-package directories
extracted from the archive.
}
\keyword{Debian}
\keyword{binary}
\keyword{extract}
\keyword{untar}
